requirements.txt
dspy-ai>=2.4.0
ollama>=0.1.7
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
ruff>=0.1.0